requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2
psycopg2-binary
PyJWT
pytz
sqlparse